git.changelog signature change and testcases
The method git.changelog takes an "options.url"-parameter instead of "options.githubUrl".
The parameter is supposed to be the repository url (no matter what provider), which will be
converted to a markdown to the commit-website. This currently only works for github-urls.
Other URLs will be ignored and no link will be rendered.